change server folder name to src
